fix: move loadCampaign back to component and drop unreachable spy test
