+++ test/components/NostoCampaign.spec.ts
@@ -2,8 +2,6 @@ import { describe, it, beforeEach, expect, vi } from "vitest"
 import { mockNostojs } from "@nosto/nosto-js/testing"
 import "@/components/NostoCampaign/NostoCampaign"
 import { NostoCampaign } from "@/components/NostoCampaign/NostoCampaign"
-import * as campaignUtils from "@/components/NostoCampaign/loadCampaign"
-
 import { RequestBuilder } from "@nosto/nosto-js/client"
 
 describe("NostoCampaign", () => {
@@ -26,25 +24,6 @@ describe("NostoCampaign", () => {
   it("should throw in connectedCallback if placement is missing", async () => {
     campaign = new NostoCampaign()
     await expect(campaign.connectedCallback()).rejects.toThrow("Property placement is required.")
-  })
-
-  it("should call loadCampaign() with placement, product, and variant", async () => {
-    const spy = vi.spyOn(campaignUtils, "loadCampaign").mockImplementation(async () => Promise.resolve())
-
-    campaign = mount({
-      placement: "789",
-      product: "123",
-      variant: "var1"
-    })
-
-    await campaign.connectedCallback()
-
-    expect(campaign.placement).toBe("789")
-    expect(campaign.product).toBe("123")
-    expect(campaign.variant).toBe("var1")
-    expect(spy).toHaveBeenCalled()
-
-    spy.mockRestore()
   })
 
   it("should mark element for client injection", async () => {
